requirements.txt
coreapi==2.3.3
Django==4.0
django-bootstrap3==21.1
djangorestframework==3.12.4
#django-oauth-toolkit==1.5.0
drf-yasg==1.20.0
Jinja2==3.0.3
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
redis==4.0.2